+++ kipper/core/src/compiler/ast/common/ast-types.ts
@@ -208,6 +208,7 @@ export type ASTExpressionKind =
 	| typeof ParseRuleKindMapping.RULE_identifierTypeSpecifierExpression
 	| typeof ParseRuleKindMapping.RULE_genericTypeSpecifierExpression
 	| typeof ParseRuleKindMapping.RULE_typeofTypeSpecifierExpression
+	| typeof ParseRuleKindMapping.RULE_nullableTypeSpecifierExpression
 	| typeof ParseRuleKindMapping.RULE_bitwiseOrExpression
 	| typeof ParseRuleKindMapping.RULE_bitwiseAndExpression
 	| typeof ParseRuleKindMapping.RULE_bitwiseXorExpression
@@ -296,6 +297,7 @@ export type ASTExpressionRuleName =
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_identifierTypeSpecifierExpression]
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_genericTypeSpecifierExpression]
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_typeofTypeSpecifierExpression]
+	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_nullableTypeSpecifierExpression]
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_bitwiseOrExpression]
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_bitwiseAndExpression]
 	| (typeof KindParseRuleMapping)[typeof ParseRuleKindMapping.RULE_bitwiseXorExpression]