+++ kipper/core/src/compiler/ast/nodes/expressions/primary-expression/object-primary-expression/object-primary-expression.ts
@@ -81,7 +81,6 @@ export class ObjectPrimaryExpression extends PrimaryExpression<
 		if (!children) {
 			throw new UnableToDetermineSemanticDataError();
 		}
-
 		const keyValuePairs = <ObjectProperty[]>this.children;
 
 		this.semanticData = {