+++ CHANGELOG.md
@@ -20,6 +20,15 @@ To use development versions of Kipper download the
 
 - New cast keywords `cast as`, `force as` and `try as`, which allow for various type-safe cast operations.
   ([#685](https://github.com/Kipper-Lang/Kipper/issues/685))
+- Implemented new `T?` and `T??` nullable type operator which marks the given type as `T | null` and `T | undefined`
+  respectively.
+- New classes:
+  - `NullableTypeSpecifierExpression`, which represents an AST nullable type specifier expression that marks a given
+    type as nullable with either `null` or `undefined`.
+- New interfaces:
+  - `NullableTypeSpecifierExpressionSemantics`, which represents the semantics of a nullable type specifier expression.
+  - `NullableTypeSpecifierExpressionTypeSemantics`, which represents the type semantics of a nullable type specifier
+    expression.
 - New functions:
   - `KipperTypeScriptTarget.getRuntimeType`, which gets the corresponding runtime representation of a type.
 - New properties:
@@ -33,6 +42,14 @@ To use development versions of Kipper download the
   - `CustomType.sourceNode`, which returns the source node of the custom type.
   - `CustomType.clsConstructor`, which returns the constructor of the class if the type is a class and the class has a
     constructor.
+- New types:
+  - `KipperNullableNullOperator`, which is the `null` nullable operator that can be used in Kipper.
+  - `KipperNullableUndefinedOperator`, which is the `undefined` nullable operator that can be used in Kipper.
+  - `KipperNullableOperators`, which are the nullable operators that can be used in Kipper.
+- New constants:
+  - `kipperNullableNullOperator`, which is the `null` nullable operator that can be used in Kipper.
+  - `kipperNullableUndefinedOperator`, which is the `undefined` nullable operator that can be used in Kipper.
+  - `kipperNullableOperators`, which are the nullable operators that can be used in Kipper.
 
 ### Changed